Rework completion
Completion is moved to a dedicated plugin.
Triggering completion is done with the Complete event with a tuple with
input, cursor and password state. Completion result is returned with
Completed event with the completed input.

Completion state is stored in the completion plugin. Stopping current
completion iteration is done with the CompletionStop event.