Use Arc<Mutex<>> for aparte
Aparte is used either directly with a ref or within an Arc<Mutex<>>.
Event loop uses an Arc<Mutex<>> because it must be able to spawn new tasks.

Command are executed directly with a mut ref to Aparte. If it needs to call a
method using an Arc<Mutex<>> it should use a proxy method that will trigger a
dedicated event. On next event loop the real method will be called with the
Arc<Mutex<>>.