Rework terminus
Don't wrap view in View and don't rely on Specification which is
Nightly.

Instead each widget must impl View, and data that is common to View is
stored in Dimension and passed to all required function.

Also remove screen ref inside View and pass it to render function.

If an event used to write to screen now it has to set it's state to
dirty and wait end of event loop so that terminus will trigger a new rendering.