Rework message handling workflow
Message can be handled by any plugin, then core won't handle it.
Handling is done with two new function in Plugin interface.
can_handle_message and handle_message, the former returns [0..1] to say
how confident it is for handling this message, the most confident is
then choosen and the latter function is called to really handle the
message. If a new message is generated (mam, carbons, encryption, ...)
it is processed by scheduling a new RawMessage event.